+++ CHANGELOG.md
@@ -33,6 +33,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Services directory
 - Category model
 
+### Updated
+
+- Filter on having order
+
 ## [0.1.0] - 2025-06-16 - Vlad
 
 ### Updated